fix: resolve timezone bug in calcDDay
